+++ test/test_content_aware_timelapse.py
@@ -1,8 +1,0 @@
-"""Tests for `content_aware_timelapse` package."""
-
-
-def test_main() -> None:
-    """
-    TODO: Test for the main entry point of content_aware_timelapse
-    :return: None
-    """